refactor(shared): parse-duration을 shared로 승격 — 구획 스킵판정 재사용 (#444)
