+++ src/shared/lib/functions/parse-duration.ts
